backend/requirements.txt
aiofiles==24.1.0
alembic==1.11.1
asyncpg==0.29.0
bcrypt==4.3.0
fastapi==0.109.2
greenlet==3.2.0
passlib==1.7.4
pillow==11.2.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.36
uvicorn==0.22.0